@@ -1,25 +1,26 @@
-const axios = require("axios").default;
+// const axios = require("axios").default;
+import axios from "axios";
 // const url = "https://art-selling-website.onrender.com"
-// const url = "http://localhost:8000";
+const url = "http://localhost:8000";
 
 export const getAllUser = async (page,limit) => {
   return await axios({
     method: "get",
-    url: `/users?page=${page}&limit=${limit}`,
+    url: `${url}/users?page=${page}&limit=${limit}`,
   });
 };
 
 export const getSomeUser = async () => {
   return await axios({
     method: "get",
-    url: `/someUsers`,
+    url: `${url}/someUsers`,
   });
 };
 
 export const suspend = async (form) => {
   return await axios({
     method: "post",
-    url: `/suspend`,
+    url: `${url}/suspend`,
     data: form,
     headers: { "Content-Type": "application/json" },
   });
