@@ -1,11 +1,12 @@
-const axios = require("axios").default;
+// const axios = require("axios").default;
+import axios from "axios";
 // const url = "https://art-selling-website.onrender.com"
-// const url = "http://localhost:8000";
+const url = "http://localhost:8000";
 
 export const addtocart = async (form) => {
   return await axios({
     method: "post",
-    url: `/addtocart`,
+    url: `${url}/addtocart`,
     data: form,
     headers: { "Content-Type": "application/json" },
   });
@@ -14,14 +15,14 @@ export const addtocart = async (form) => {
 export const mycart = async (id) => {
   return await axios({
     method: "get",
-    url: `/mycart/${id}`,
+    url: `${url}/mycart/${id}`,
   });
 };
 
 export const updateCart = async (form) => {
   return await axios({
     method: "post",
-    url: `/updateCart`,
+    url: `${url}/updateCart`,
     data: form,
     headers: { "Content-Type": "application/json" },
   });
@@ -31,7 +32,7 @@ export const deleteItem = async (form) => {
   console.log(form);
   return await axios({
     method: "delete",
-    url: `/deleteItem`,
+    url: `${url}/deleteItem`,
     data: form,
     headers: { "Content-Type": "application/json" },
   });
@@ -41,7 +42,7 @@ export const buyCart = async (form) => {
   console.log(form);
   return await axios({
     method: "post",
-    url: `/buyCart`,
+    url: `${url}/buyCart`,
     data: form,
     headers: { "Content-Type": "application/json" },
   });
@@ -50,28 +51,28 @@ export const buyCart = async (form) => {
 export const getTrack = async () => {
   return await axios({
     method: "get",
-    url: `/getTrack`,
+    url: `${url}/getTrack`,
   });
 };
 
 export const getAcceptedTrack = async () => {
   return await axios({
     method: "get",
-    url: `/getAcceptedTrack`,
+    url: `${url}/getAcceptedTrack`,
   });
 };
 
 export const getDeliveredTrack = async () => {
   return await axios({
     method: "get",
-    url: `/getDeliveredTrack`,
+    url: `${url}/getDeliveredTrack`,
   });
 };
 
 export const isDelivered = async (form) => {
   return await axios({
     method: "post",
-    url: `/isDelivered`,
+    url: `${url}/isDelivered`,
     data: form,
     headers: { "Content-Type": "application/json" },
   });
@@ -80,7 +81,7 @@ export const isDelivered = async (form) => {
 export const isNotDelivered = async (form) => {
   return await axios({
     method: "post",
-    url: `/isNotDelivered`,
+    url: `${url}/isNotDelivered`,
     data: form,
     headers: { "Content-Type": "application/json" },
   });
@@ -89,7 +90,7 @@ export const isNotDelivered = async (form) => {
 export const isAccepted = async (form) => {
   return await axios({
     method: "post",
-    url: `/isAccepted`,
+    url: `${url}/isAccepted`,
     data: form,
     headers: { "Content-Type": "application/json" },
   });
