@@ -1,11 +1,12 @@
-const axios = require("axios").default;
+// const axios = require("axios").default;
+import axios from "axios";
 // const url = "https://art-selling-website.onrender.com"
-// const url = "http://localhost:8000";
+const url = "http://localhost:8000";
 
 export const register = async (form) => {
   return await axios({
     method: "post",
-    url: `/register`,
+    url: `${url}/register`,
     data: form,
     headers: { "Content-Type": "application/json" },
   });
@@ -14,7 +15,7 @@ export const register = async (form) => {
 export const login = async (form) => {
   return await axios({
     method: "post",
-    url: `/login`,
+    url: `${url}/login`,
     data: form,
     headers: { "Content-Type": "application/json" },
   });
@@ -23,7 +24,7 @@ export const login = async (form) => {
 export const googleLogin = (googleData) => {
   return axios({
     method: "POST",
-    url: `/googleLogin`,
+    url: `${url}/googleLogin`,
     data: { token: googleData.tokenId },
     headers: {
       "Content-Type": "application/json",
@@ -34,7 +35,7 @@ export const googleLogin = (googleData) => {
 export const autho = () => {
   return axios({
     method: "get",
-    url: `/auth`,
+    url: `${url}/auth`,
     // headers : {"Content-Type": "application/json"}
   });
 };
@@ -42,14 +43,14 @@ export const autho = () => {
 export const refreshRoute = () => {
   return axios({
     method: "post",
-    url: `/refreshToken`,
+    url: `${url}/refreshToken`,
   });
 };
 
 export const update = async (form) => {
   return await axios({
     method: "post",
-    url: `/update`,
+    url: `${url}/update`,
     data: form,
     headers: { "Content-Type": "multipart/form-data" },
   });
@@ -58,16 +59,16 @@ export const update = async (form) => {
 export const getProfile = async (id) => {
   return await axios({
     method: "get",
-    url: `/profile/${id}`,
+    url: `${url}/profile/${id}`,
     // data: form,
-    // headers: { "Content-Type": "application/json" },
+    headers: { "Content-Type": "application/json" },
   });
 };
 
 export const post = async (form) => {
   return await axios({
     method: "post",
-    url: `/postArt`,
+    url: `${url}/postArt`,
     data: form,
     headers: { "Content-Type": "multipart/form-data" },
   });
@@ -76,21 +77,21 @@ export const post = async (form) => {
 export const getPost = async (page, limit) => {
   return await axios({
     method: "get",
-    url: `/getPosts?page=${page}&limit=${limit}`,
+    url: `${url}/getPosts?page=${page}&limit=${limit}`,
   });
 };
 
 export const getSomePosts = async () => {
   return await axios({
     method: "get",
-    url: `/getSomePosts`,
+    url: `${url}/getSomePosts`,
   });
 };
 
 export const deletePost = async (form) => {
   return await axios({
     method: "delete",
-    url: `/deletePost`,
+    url: `${url}/deletePost`,
     data: form,
     headers: { "Content-Type": "application/json" },
   });
@@ -99,14 +100,14 @@ export const deletePost = async (form) => {
 export const getAPost = async (id) => {
   return await axios({
     method: "get",
-    url: `/getAPost/${id}`,
+    url: `${url}/getAPost/${id}`,
   });
 };
 
 export const updatePost = async (form) => {
   return await axios({
     method: "put",
-    url: `/updatePost`,
+    url: `${url}/updatePost`,
     data: form,
     headers: { "Content-Type": "multipart/form-data" },
   });
@@ -115,28 +116,28 @@ export const updatePost = async (form) => {
 export const getMyPosts = async (page, limit) => {
   return await axios({
     method: "get",
-    url: `/getMyPosts?page=${page}&limit=${limit}`,
+    url: `${url}/getMyPosts?page=${page}&limit=${limit}`,
   });
 };
 
 export const getBoughtItems = async (page, limit) => {
   return await axios({
     method: "get",
-    url: `/getBoughtItems?page=${page}&limit=${limit}`,
+    url: `${url}/getBoughtItems?page=${page}&limit=${limit}`,
   });
 };
 
 export const getUsersPosts = async (id,page,limit) => {
   return await axios({
     method: "get",
-    url: `/getUsersPosts/${id}?page=${page}&limit=${limit}`,
+    url: `${url}/getUsersPosts/${id}?page=${page}&limit=${limit}`,
   });
 };
 
 export const addWallet = async (form) => {
   return await axios({
     method: "post",
-    url: `/addWallet`,
+    url: `${url}/addWallet`,
     data: form,
     headers: { "Content-Type": "application/json" },
   });
